(core) restructure write block
