+++ generators/client-2/templates/_.bowerrc
