+++ generators/client-2/templates/_.eslintignore
