+++ generators/client-2/templates/_.eslintrc.json
